Refactor cmd to multiple files. Opti delete cmd to use bulk delete. Add new option on list cmd.
