+++ pkg/message/message.go
@@ -13,10 +13,5 @@ func List(s *discordgo.Session, channelID string, limit int, beforeID, afterID, 
 		log.With("err", err).Error("can't list message")
 		return messageList, errors.New("can't list message")
 	}
-
-	if len(messageList) == 0 {
-		return messageList, errors.New("no message")
-	}
-
 	return messageList, nil
 }
